Fix EF configuration for Id property of entities
HasKey assumes that the value is generated by the DB, but we set the ID in code. This causes problem for the collection of time records which are not explicitly added with db.TimeRecords.Add(task). Because of the ID, EF treats a new TimeRecord as Modified and tries to UPDATE it instead of INSERTing it.

https://learn.microsoft.com/en-us/ef/core/modeling/generated-properties?tabs=fluent-api